requirements.txt
pyzmq
requests
numpy
pyyaml
